whoismrsaemir/requirements.txt
beautifulsoup4==4.7.1
future==0.17.1
lxml==4.3.1
python-whois==0.7.1
whois==0.7
Django==2.1.2
djangorestframework==3.9.0
jdatetime==3.2.0
